+++ lib/kamal_gitlab_review_app/naming.rb
@@ -6,14 +6,16 @@ module KamalGitlabReviewApp
       def for_mr(iid:)
         iid_str = normalize_iid(iid)
         service = "#{KamalGitlabReviewApp::Env.service_prefix}_#{iid_str}"
-
-        {
+        names = {
           environment_name: "#{KamalGitlabReviewApp::Env.environment_prefix}-#{iid_str}",
           host: "#{KamalGitlabReviewApp::Env.host_label_prefix}-#{iid_str}.#{KamalGitlabReviewApp::Env.review_domain}",
-          service: service,
-          # Kamal accessory container name is "#{service}-#{accessory}" (see Kamal::Configuration::Accessory#service_name).
-          db_host: "#{service}-db"
+          service: service
         }
+
+        # Kamal accessory container name is "#{service}-#{accessory}" (see Kamal::Configuration::Accessory#service_name).
+        db_accessory = KamalGitlabReviewApp::Env.db_accessory
+        names[:db_host] = "#{service}-#{db_accessory}" if db_accessory
+        names
       end
 
       private