+++ spec/dns/cloudflare_spec.rb
@@ -64,6 +64,50 @@ RSpec.describe KamalGitlabReviewApp::Dns::Cloudflare do
       expect { provider.upsert_a_record!(name:, ip:, ttl: 120) }
         .to raise_error(KamalGitlabReviewApp::Dns::Error, /Cloudflare request failed/)
     end
+
+    it 'raises Dns::Error when the response body is not JSON' do
+      stub_request(:get, %r{\Ahttps://api\.cloudflare\.com/client/v4/zones/zone123/dns_records})
+        .to_return(status: 502, body: '<html>Bad Gateway</html>')
+
+      expect { provider.upsert_a_record!(name:, ip:, ttl: 120) }
+        .to raise_error(KamalGitlabReviewApp::Dns::Error, /Cloudflare request failed/)
+    end
+
+    it 'raises Dns::Error when HTTP status is not success even if body looks ok' do
+      stub_request(:get, %r{\Ahttps://api\.cloudflare\.com/client/v4/zones/zone123/dns_records})
+        .to_return(status: 500, body: { success: true, result: [] }.to_json)
+
+      expect { provider.upsert_a_record!(name:, ip:, ttl: 120) }
+        .to raise_error(KamalGitlabReviewApp::Dns::Error, /HTTP 500/)
+    end
+
+    it 'configures open and read timeouts on the HTTP client' do
+      http = instance_double(Net::HTTP)
+      allow(Net::HTTP).to receive(:new).and_return(http)
+      allow(http).to receive(:use_ssl=)
+      allow(http).to receive(:open_timeout=)
+      allow(http).to receive(:read_timeout=)
+      allow(http).to receive(:request).and_return(
+        instance_double(Net::HTTPResponse, code: '200', body: { success: true, result: [] }.to_json)
+      )
+
+      provider.upsert_a_record!(name:, ip:, ttl: 120)
+
+      expect(http).to have_received(:open_timeout=).with(described_class::OPEN_TIMEOUT).at_least(:once)
+      expect(http).to have_received(:read_timeout=).with(described_class::READ_TIMEOUT).at_least(:once)
+    end
+
+    it 'wraps network timeouts as Dns::Error' do
+      http = instance_double(Net::HTTP)
+      allow(Net::HTTP).to receive(:new).and_return(http)
+      allow(http).to receive(:use_ssl=)
+      allow(http).to receive(:open_timeout=)
+      allow(http).to receive(:read_timeout=)
+      allow(http).to receive(:request).and_raise(Net::OpenTimeout)
+
+      expect { provider.upsert_a_record!(name:, ip:, ttl: 120) }
+        .to raise_error(KamalGitlabReviewApp::Dns::Error, /timed out|timeout/i)
+    end
   end
 
   describe '#delete_record!' do